Revert "Fixed, refactored, covered with tests and proper typings for batching (#1488)"
This reverts commit ee00b19db135b1e3ebf86547c449dc6455f6a7d1.
