+++ src/neptune/internal/operation.py
@@ -19,7 +19,6 @@ from dataclasses import dataclass
 from datetime import datetime
 from typing import (
     TYPE_CHECKING,
-    Any,
     Generic,
     List,
     Optional,
@@ -295,13 +294,7 @@ class UploadFileSet(Operation):
 class LogOperation(Operation, abc.ABC):
     @abc.abstractmethod
     def value_count(self) -> int:
-        ...
-
-    # Workaround for mypy and not being able to combine dataclasses and abstract properties with subclass
-    #  initialization in series
-    @abc.abstractmethod
-    def get_values(self) -> List[Any]:
-        ...
+        pass
 
 
 @dataclass
@@ -344,9 +337,6 @@ class LogFloats(LogOperation):
     def value_count(self) -> int:
         return len(self.values)
 
-    def get_values(self) -> List[LogSeriesValue[float]]:
-        return self.values
-
 
 @dataclass
 class LogStrings(LogOperation):
@@ -372,9 +362,6 @@ class LogStrings(LogOperation):
 
     def value_count(self) -> int:
         return len(self.values)
-
-    def get_values(self) -> List[LogSeriesValue[str]]:
-        return self.values
 
 
 @dataclass
@@ -423,9 +410,6 @@ class LogImages(LogOperation):
 
     def value_count(self) -> int:
         return len(self.values)
-
-    def get_values(self) -> List[LogSeriesValue[ImageValue]]:
-        return self.values
 
 
 @dataclass
