+++ src/neptune/internal/queue/disk_queue.py
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-__all__ = ["DiskQueue", "QueueElement"]
+__all__ = ["DiskQueue"]
 
 import json
 import logging
